Move global state context, reducer and hooks into store folder
