+++ src/App.js
@@ -1,5 +1,6 @@
-import { GlobalProvider } from "./GlobalState";
 import UserDashboard from "./views/UserDashboard";
+
+import { GlobalProvider } from "./store/GlobalState";
 
 const App = () => (
   <GlobalProvider>
